+++ tests/test_tensor_product.py
@@ -46,7 +46,7 @@ class TestTensorProductImport:
             instructions,
             shared_weights=shared_weights,
             internal_weights=internal_weights,
-        ).to(torch.float32)
+        )
 
         self._assert_forward_matches(
             torch_module=torch_module,
@@ -83,7 +83,7 @@ class TestTensorProductImport:
             shared_weights=shared_weights,
             internal_weights=internal_weights,
             method='naive',
-        ).to(torch.float32)
+        )
 
         self._assert_forward_matches(
             torch_module=torch_module,
@@ -116,8 +116,8 @@ class TestTensorProductImport:
         dim_in1 = jax_irreps_in1.dim
         dim_in2 = jax_irreps_in2.dim
 
-        x1_np = self.rng.standard_normal((self.batch, dim_in1)).astype(np.float32)
-        x2_np = self.rng.standard_normal((self.batch, dim_in2)).astype(np.float32)
+        x1_np = self.rng.standard_normal((self.batch, dim_in1))
+        x2_np = self.rng.standard_normal((self.batch, dim_in2))
 
         x1_jax = jnp.array(x1_np)
         x2_jax = jnp.array(x2_np)
@@ -146,17 +146,11 @@ class TestTensorProductImport:
         else:
             if shared_weights:
                 if backend == 'cue':
-                    weights_np = self.rng.standard_normal((1, weight_numel)).astype(
-                        np.float32
-                    )
+                    weights_np = self.rng.standard_normal((1, weight_numel))
                 else:
-                    weights_np = self.rng.standard_normal((weight_numel,)).astype(
-                        np.float32
-                    )
+                    weights_np = self.rng.standard_normal((weight_numel,))
             else:
-                weights_np = self.rng.standard_normal(
-                    (self.batch, weight_numel)
-                ).astype(np.float32)
+                weights_np = self.rng.standard_normal((self.batch, weight_numel))
 
             weights_jax = jnp.array(weights_np)
             weights_torch = torch.tensor(weights_np)
@@ -176,8 +170,8 @@ class TestTensorProductImport:
             params = transformed.init(key, x1_jax, x2_jax)
             out_jax = transformed.apply(params, None, x1_jax, x2_jax)
 
-        x1_torch = torch.tensor(x1_np, dtype=torch.float32)
-        x2_torch = torch.tensor(x2_np, dtype=torch.float32)
+        x1_torch = torch.tensor(x1_np)
+        x2_torch = torch.tensor(x2_np)
         with torch.no_grad():
             if internal_weights:
                 out_torch = torch_module(x1_torch, x2_torch).cpu().numpy()
